Moving packages and fix reference
